driver: cleanup, merge `common.hpp` & `generic.hpp`
